Add a vote keeper (#4)
* Add a vote keeper

* Rename `polka_round` to `pol_round` (Proof-Of-Lock)

* Add unit tests to VoteKeeper

* Add codecov settings

* Add tests for precommit
