+++ Code/QUESTIONS.md
@@ -1,0 +1,2 @@
+- How do we deal with errors?
+- How do we parametrize over values, id(v), etc.
