+++ Code/TODO.md
@@ -1,0 +1,13 @@
+is proposal complete
+if polka not nil, then need to see proof of lock (2f+1 votes)
+then send proposal
+
+count votes for cur, prev, 1/2 next round
+
+if complete proposal from a past round => to current one
+if we have some threshold (f+1) of votes for a future round => skip to that round
+
+context (get proposer, get value)
+signing contextt
+
+abstract over values and validator sets
