+++ Code/round/Cargo.toml
@@ -3,7 +3,5 @@ name    = "malachite-round"
 version = "0.1.0"
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
 malachite-common = { version = "0.1.0", path = "../common" }
