+++ codecov.yml
@@ -1,0 +1,12 @@
+codecov:
+  require_ci_to_pass: yes
+
+coverage:
+  precision: 2
+  round: nearest
+  range: "50...100"
+
+  status:
+    project: true
+    patch: true
+    changes: true
